+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/StaticMetadataManagerTest.java
@@ -128,9 +128,6 @@ public class StaticMetadataManagerTest {
 		Assert.assertEquals("v1", metadata.get("k1"));
 		Assert.assertEquals("v22", metadata.get("k2"));
 		Assert.assertEquals("v33", metadata.get("k3"));
-		Assert.assertEquals("zone2", metadata.get("zone"));
-		Assert.assertEquals("region1", metadata.get("region"));
-		Assert.assertEquals("campus1", metadata.get("campus"));
 
 		Map<String, String> transitiveMetadata = metadataManager.getMergedStaticTransitiveMetadata();
 		Assert.assertEquals(2, transitiveMetadata.size());