+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/config/RpcEnhancementAutoConfiguration.java
@@ -17,6 +17,7 @@
 
 package com.tencent.cloud.rpc.enhancement.config;
 
+import java.util.Collections;
 import java.util.List;
 
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
@@ -30,13 +31,14 @@ import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateReport
 import com.tencent.cloud.rpc.enhancement.resttemplate.PolarisResponseErrorHandler;
 import com.tencent.polaris.api.core.ConsumerAPI;
 
+import org.springframework.beans.factory.SmartInitializingSingleton;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.client.loadbalancer.RestTemplateCustomizer;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -100,6 +102,10 @@ public class RpcEnhancementAutoConfiguration {
 	@ConditionalOnClass(name = "org.springframework.web.client.RestTemplate")
 	protected static class PolarisRestTemplateAutoConfiguration {
 
+		@LoadBalanced
+		@Autowired(required = false)
+		private List<RestTemplate> restTemplates = Collections.emptyList();
+
 		@Bean
 		public EnhancedRestTemplateReporter polarisRestTemplateResponseErrorHandler(
 				RpcEnhancementReporterProperties properties, ConsumerAPI consumerAPI,
@@ -108,8 +114,12 @@ public class RpcEnhancementAutoConfiguration {
 		}
 
 		@Bean
-		public RestTemplateCustomizer setErrorHandlerCustomizer(EnhancedRestTemplateReporter reporter) {
-			return restTemplate -> restTemplate.setErrorHandler(reporter);
+		public SmartInitializingSingleton setErrorHandlerForRestTemplate(EnhancedRestTemplateReporter reporter) {
+			return () -> {
+				for (RestTemplate restTemplate : restTemplates) {
+					restTemplate.setErrorHandler(reporter);
+				}
+			};
 		}
 	}
 }